rgder

+++ CPP09/ex02/PmergeMe.hpp
@@ -8,6 +8,7 @@
 #include <cstdlib>
 #include <iomanip>
 #include <iterator>
+#include <utility>
 
 class PmergeMe
 {
@@ -17,11 +18,11 @@ class PmergeMe
 
         void exec(int argc, char **argv);
 
-        void recursiveDeque(std::deque<std::pair<int, int>> &pairList);
-        void recursiveVector(std::vector<std::pair<int, int>> &pairList);
+        void recursiveDeque(std::deque<std::pair<int, int> > &pairList);
+        void recursiveVector(std::vector<std::pair<int, int> > &pairList);
         int jacobsthal(int n);
-        void binarySearchDeque(int half, int n);
-        void binarySearchVector(int half, int n);
+        void binarySearchDeque(size_t half, int n);
+        void binarySearchVector(size_t half, int n);
         clock_t sortDeque(int argc, char **argv);
         clock_t sortVector(int argc, char **argv);
         std::deque<int> dequeList;
